BF(+skip): r- path strings, skip test_asyncio_loop_noninterference1 on windows for python < 3.8
First it was failing because paths strings on windows contain \ so I had
to announce those strings as r(aw).

But then test still fails since apparently on windows (with 3.7, not
3.8) there is no event loop for some reason, and get_event_loop() would
just raise RuntimeError.  Since "tuning" the reproducer might actually
change the use case we are trying to test (which works on other
environements), I decided to just announce it a skip.