TST: Recompute dynamic common_cfg defaults when testing
setup_package() overrides $HOME during the tests.  This invalidates
the default values for a good number of configuration options.  One
particularly problematic case is datalad.locations.sockets because
tests inspect SSHManager's socket_dir, which still points to the
non-test location (e.g., $HOME/.cache/datalad/sockets/), _but_ when
'datalad sshrun' is executed via git or git-annex, the test location
is in effect.

test_annex_ssh() is the test that should hit into the above scenario.
The reason it doesn't fail is because git-annex isn't actually going
through datalad-sshrun despite GIT_SSH_COMMAND being set, so
everything is operating with the non-test location.  That will change
with the next commit.

To fix this, store the defaults that are subject to change as
functions and add a helper that uses these to set "default" in
common_cfg.definitions.  This gives setup_package() and
teardown_package() an easy hook to use.