BF: Remove failed attempt to clone specific SHA
When we fail to retrieve the commit we were asked to clone, remove
failed clone to allow for trying next URL candidate.
Apart from not allowing to proceed otherwise, we'd potentially not only
leave an unwanted repo but potentially a modified submodule which in
turn may lead to infinite recursion w/ `datalad install -r`.

(Closes #5387)