BF: log: Add target name filters to ProgressHandler
get_initialized_logger() used to add two handlers to the logger for
interactive calls:

  * a plain log handler with a filter to remove dlm_progress records
  * a ProgressHandler with a filter to keep dlm_progress records

Since 8f776d396e (ENH: make ProgressHandler emit via default handler
if not a progress record, 2020-10-12), only ProgressHandler is added
as a handler, and its .emit() relays non-dlm_progress records to the
plain handler.

This change causes `test_log.test_filters` to fail when executed with
--nocapture:

    AssertionError: 'log3' unexpectedly found in
    '[DEBUG] log1\n[INFO] log2\n[INFO] log3\n

The non-test aspect to the failure is that the target name filter that
is added to the plain logger isn't honored in the new arrangement.
Fix that by adding those filters to ProgressHandler's (currently
empty) .filters.

An additional issue is that swallow_logs() discards filters from
ProgressHandler when setting up its custom handler.  That made sense
before 8f776d396e, when ProgressHandler had a filter that only kept
dlm_progress records, but it doesn't work if ProgressHandler is used
as the primary handler that relays to the plain one.

Re: #5402