RF: push - get diff depth-first not breadth-first
Remote repositories must exist for push already, and might have
hooks enabled to process the "clean" state of the repository.
In hooks we deploy with web-ui e.g. we "aggregate" information
about sizes within submodules.  For that to work correctly,
submodules should be pushed first, and not after their
"superdatasets".

Closes #5410