ENH: AnnexJsonProtocol: Add file name to progress bar label
AnnexJsonProtocol uses the command from the action result as the
progress bar label.  When operations act on a particular file and work
in parallel, there's no indication about which file a progress bar is
for:

  Get:  15%...
  Get:  25%...

Prior to 0.14.0 (in particular 46d78a8353), when AnnexRepo.get()
didn't use AnnexJsonProtocol, the above would be displayed as

  foo:  15%...
  foobar:  25%...

Update AnnexJsonProtocol to include the file name, if present,
effectively merging the above two displays into

  Get foo:  15%...
  Get foobar:  25%...

The display is a bit uglier because the progress bars aren't aligned
unless the files happen to be the same length and because long file
names can take up a lot of space, but both of those seem better than
not having the information at all.  The latter will be mitigated by
the next commit.