Deprecate old ORA configs
Configuring ssh-host and base-path for an ORA remote predates ria+ URLs
and the consideration of git's URL rewrites. Those configs are also not
capable of specifying an HTTP connection let alone a separation of `url`
and `push-url`. Respect them for now if we don't find a URL, but issue a
deprecation warning. Will be removed entirely eventually.