NF: cfg_noannex procedure to convert annexed repo into the one without annex
Initial motivation is to be able to use  datalad create -c noannex ,
in particular in the light of https://github.com/datalad/datalad/issues/5423
RFC.

Yes, we do have  --no-annex  option but it could not be applied "post-fact".
Also I like the consistency of having dataset configuration via  -c option
to create.