BF: AnnexJsonProtocol: Process final non-action record
_proc_json_record() considers only byte-progress records when updating
individual progress bars.  However, --json-progress output from
git-annex commands may not include a progress record where the
byte-progress value equals the total size.  The only indication that
the record is complete, then, is the final non-progress record for the
operation:

  {"byte-progress":14173395,
    "action":{"command":"get",
              "key":"SHA256E-s25194376--...",
              "file":"foo",
              ...},
    "total-size":25194376,
    "percent-progress":"56.26%"}
  {"command":"get","note":"from origin...\nchecksum...","success":true,
   "key":"SHA256E-s25194376--...",
   "file":"foo",
    ...}

Rework _proc_json_record() so that the final record triggers the
progress bar tear down.  In order to process a bar's completion, the
final non-progress record needs to be linked to the byte-progress bar.
While the action record seems to be a reliable link across progress
records, it isn't present in the final non-progress record, so hard
code a set of fields---command, key, and file---that should be
sufficient for linking the records for a single operations.

This new approach also fixes an error in the progress ID logic.  The
old approach was to generate a progress bar ID for incoming data by
hashing id(self) and the action of the json record.  However, the
calculation only considered the action's _keys_.  That means that it
would, for example, incorrectly collapse get() results for different
files into the same progress bar because the action results share the
same set of keys.

Fixes #5419.