ENH: AnnexJsonProtocol: Truncate long file names
In 0.14 commands using ProcessAnnexProgressIndicators were switched to
AnnexJsonProtocol, and ProcessAnnexProgressIndicators was removed
(948ccf3e18).  This switch dropped the file names from the progress
bars, and the last commit added them back.  Long file names, though,
can interfere with the display of the progress bar.

Deal with this by porting similar truncation logic to what was in the
now-removed ProcessAnnexProgressIndicators.