RF: AnnexJsonProtocol: Always log action error messages
_proc_json_record() logs error messages from the _action_ value [*].
However, it only does so when "byte-progress" is in the top-level
record.  I'm not sure if there's ever a case where the "action" and
"byte-progress" keys aren't both present, but, since the current code
bothers to check both conditions, move the logging of the errors
upstream so that it's triggered even if byte-progress isn't present.

[*] It's not clear to me that git-annex ever adds error messages to
    this value instead of the top-level record.