TST: assert_not_in_results: Make failure message more helpful
When assert_not_in_results() fails, it raises an assertion without a
message.  On the other hand, assert_in_results() more helpfully
displays the query and the set of results searched.  Rewrite
assert_in_results() and assert_not_in_results() to use a common helper
function, extracting the logic is relevant for both.

The end result is that assert_not_in_results() shows a failure message
like this:

  AssertionError: Result
    {
     "action": "unlock",
    }
  unexpectedly found among
    [
     {
      "action": "unlock",
      ...
     }
    ]