TST: Disable non-decorator appveyor skips
Remove hidden test skips (hidden in that they are not declared with the
test itself) in test helpers, that are merely based on running on
AppVeyor. This was originally addressing issues with the test setup for
Windows. Since we switched to test several platforms on AppVeyor and the
new setup works with more tests now, remove those and subsequently be
more specific where skips still are necessary.