BF: gitrepo: Honor _GIT_COMMON_OPTIONS more widely
GitRepo's _GIT_COMMON_OPTIONS is used in _call_git(), _call_annex(),
and AnnexRepo._batched, but several places in GitRepo construct a git
command without _GIT_COMMON_OPTIONS.  This doesn't matter at the
moment because it's an empty list, but it will once an upcoming commit
adds an option.

Audit the git calls in GitRepo, adding _GIT_COMMON_OPTIONS to them.
Also move _GIT_COMMON_OPTIONS to a class attribute so that it can be
used in the class method clone().