BF: gitrepo: Call git with '-c diff.ignoreSubmodules=none'
If a user has configured diff.ignoreSubmodules to a value other than
the default "none", get_content_info() and friends will see the
submodule changes (as intended and expected by the code) because the
low-level commands used are not affected by diff.ignoreSubmodules.
However, a save_() call will fail at the git-commit step when the only
modifications are those that would be ignored according to the
diff.ignoreSubmodules setting.

Configure diff.ignoreSubmodules=none for all our git calls because
ignoring submodules isn't something we currently account for and
probably not something we want to support.

Fixes #3462.