RF: Make ORA's HTTP backend not a special case
`HTTPRemoteIO` now behaves largely just like SSH and local access,
except for being read-only. Remove special casing throughout
`RIARemote`. This means, that RIA stores are validated by the ORA
special remote over HTTP the same way as over SSH or local access.
Therefore providing the ria-layout-version files is now mandatory for
HTTP stores as well.

This patch should streamline things a bit and closes #5451. Thus making
the reconfiguration business in `datalad clone` consistent across
protocols. Furthermore, since the version files are now requested and
used to construct the URLs, we can simplify server-side scripts dealing
wiht archives, etc. There's also hope that this way we can allow for a
version file to define a layout directly, instead of declaring a label
that needs to be known to released datalad, because all the version is
about is locations associated with the .git, annex/objects and archives
of datasets in the store. In theory these locations could be declared in
a JSON-record or alike in that version file, allowing store maintainers
to define their thing freely.