TST: test_annex_backends: Use subdirectories for test cases
test_annex_backends() tests the configured backend when AnnexRepo is
initialized without and with the `backend` keyword argument, removing
the top-level repo path in between the two cases.

Instead create dedicated repos under it the top-level path.  It's a
matter of taste, though it arguably makes it easier to identify the
distinct test cases and to tack on new test cases (as will be done
next).