BF: annexrepo: Point default_backends to preferred option
The default_backends property only considers the annex.backends value
in the git configuration, but, as of 4c1e3210f (annex.backend is the
new name for what was annex.backends, 2017-05-09), git-annex gives
priority to annex.backend (no 's').  If annex.backend is not in
.gitattributes (preferred) or the git configuration, git-annex takes
the first backend from annex.backends if it is the git configuration.