CLN: annexrepo: Drop a to-do comment
Before git-annex 6.20170510 [*], git-annex used to try the backends
listed in annex.backends in order.  Now the annex.backend (no 's')
option is preferred.  If annex.backend isn't set in .gitattributes or
the git configuration, git-annex looks at annex.backends in the git
configuration, only considering the first value.

So, the to-do comment in set_default_backend() about how to support a
list is stale; DataLad's minimum version of git-annex doesn't support
a list.

[*] 4c1e3210f (annex.backend is the new name for what was
    annex.backends, 2017-05-09)