ENH: update: Add minimal fetching --follow=parentds mode
update(..., recursive=True) calls git-fetch in every subdataset.  When
using follow=parentds to follow the commit registered in the parent
dataset, this can lead to many unneeded fetches, especially when only
a few subdatasets need to be updated.

Add another --follow value, "parentds-lazy", that is like "parent-ds"
but fetches only if the commit isn't already present and avoids
processing the dataset entirely if the registered commit is already
checked out.  Keep "parentds" around because, as mentioned in gh-4452,
the caller may still want to fetch from the sibling to, e.g., bring in
updates to the git-annex branch.

Note that in cases where the registered commit is already checked out
in a subdataset and the subdataset is unmodified, we could
conceptually skip descending into deeper subdatasets.  However, that
would require either extending subdatasets() or reworking/exposing
internal subdatasets() logic for update() to use.

Closes #4452.
Closes #4642.