Make `--output-format default` accessible
Running something like `datalad -f default status` does not actually
switch to the standard result renderer, because we use the "tailored"
one by default. As pointed out in #2481 this not only conflates two
different meanings of *default*, but also doesn't really make sense.

From my POV the implementation really wants `"tailored"` to be the default
mode, and not the "default" renderer to be enabled by default.

Instead of getting creative with new synonyms for these terms, I went
for the minimal change that achieves the desired result (make it
possible to have a command with a custom result renderer run with the
standard one instead).

I switch the default value of `common_output_format` to `"tailored"`,
which already is the effective default behavior. This makes a
specification of an alternative `"default"` value actually switch the
renderer to the default one.

One could argue that `"default"` would be better named `"standard"`
(renderer), but I will not walk on this battle field ;-)

Fixes gh-2481