MNT: create: Drop custom result renderer
Except for create(), every command with a custom result renderer sets
the attribute result_renderer to "tailored" so that the custom
renderer is used by default.  From the command line, the default value
is now "tailored", falling back to "default" if the command class's
result_renderer isn't set.  This means that, as of the last commit,
the custom renderer of create() cannot be accessed from the command
line with `--output-format=tailored`.

Unless there are larger reworks that fix the situation, it seems there
is little point in keeping the custom renderer of create() around but
_not_ setting Create.result_renderer to "tailored".  However, an
argument against setting Create.result_renderer to "tailored" is that
current behavior for the custom renderer is confusing, likely
indicating nobody has really used it.  For example:

  $ datalad create -d. sub
  Nothing was created
  Nothing was created
  Nothing was created
  Created dataset at /tmp/dl-Yof8LkL/sub.

All those "Nothing was created" messages are from the custom render
acting on non-create results.  (It looks like it incorrectly assumes
that the result_filter removes these, but the result filter doesn't
prevent results from getting to the renderer.)