Leave a log trace which env vars are queried for credentials
Before one would need to spend 10-15min to discover how exactly this is
working. Now `--log-level 5` will give a hint:

```
[Level 5] Credentials lookup attempt DATALAD_webdav.4shared.com_password
[Level 5] Credentials lookup attempt DATALAD_webdav.4shared.com_user
```

Should not hurt anyone. Isn't beautiful either, but having to override
credentials is a relatively frequent task in development, but also in
environments without credentials stores (HPC...). Knowing the names of
the variables is key here.