BF: run: Don't remove leading empty directory when removing outputs
If outputs aren't present (and thus can't be unlocked), run() removes
them so that the command doesn't need to provide special handling of
annex links.  This is done with `datalad remove`, which calls git-rm
underneath. `git rm path/to/file` deletes the leading directories when
they are otherwise empty.

That behavior can be surprising for commands that don't ensure
outputs' directories exist.  Instead unlink the file directly so
leading empty directories stick around.  Unlike git-rm, this doesn't
remove the file from the index, but the end result is the same after
the save (even if the command ends up not recreating the output).

Note, however, that ideally commands should be written to not assume
that the output directory exists in HEAD's tree.  This makes them
better suited to be used outside the context of the current tree
(e.g., `rerun --onto`), where an output's directory of course may be
pruned because the checked out tree doesn't include any tracked files
under an output's directory.

Closes #5486.