BF: AnnexJsonProtocol - memoize the last line which failed to json load
Protocol's buffer could get filled up before the entire json line is
actually received, so the AnnexJsonProtocol could then receive only a
part of it, and fail to parse such a partial json entry.

The patch is a bit overdone: in principle I think

    not data.endswith(os.linesep.encode())

condition could be dropped, but I think it should not hurt since we are not
expecting new lines in the output within records, thus if the trailing symbol
in the entire data is a newline, it would mean that we did read the entire line
and thus we should not try to join it with the next one.

I am not sure if similar "I care about lines" behavior is relevant to other
WitlessProtocols and might need to be adjusted similarly.

Closes #5393