MNT: annexrepo: Bump minimum git-annex version to 8.20200309
Increasing GIT_ANNEX_MIN_VERSION is long overdue.  The main question
is what the new minimum should be.  Some landmarks:

  * 7.20190912 is when v5 repos started automatically upgrading (to
    v7).

  * buster-backports has 8.20200330.

  * Neurodebian currently has a recent release (8.20210223).

  * 8.20200330 is required so that subdataset changes are synced back
    to an adjusted branch (#4292).  subdataset deletions require
    8.20210127 (#5290).

  * 7.20200202.7 until 8.20200309 is incompatible with the ora special
    remote (gh-4509).

As suggested by gh-5399, we probably shouldn't increase this beyond
what's available from buster-backports.  That's 8.20200330 and would
be a good minimum due to the subdataset/adjusted branch fixes.

However, conda-forge doesn't have 8.20200330, which makes testing the
minimum on the CI a little more complicated.  It does have one release
earlier, 8.20200309.  So, go with 8.20200309 for ease of testing.

Closes #5399.