RM: SafeConfigParserWithIncludes
I could not determine where and how the functionality added by this
class over `configparser.ConfigParser` is needed. It used to be used
in datalad.config, but that was loooong ago. This change simply
removes the class (no tests to remove, because there are none), and
replaces its usage with `configparser.ConfigParser`.

Closes gh-5513