TST: travis: Move git-annex=8.20210224 builds to 8.20210310
After the previous commit switched to 8.20210224, the Travis build had
two test failures where the underlying git-annex command had an error:

  datalad.customremotes.tests.test_archives:test_annex_get_from_subdir
  datalad.customremotes.tests.test_archives:test_basic_scenario

  <https://travis-ci.com/github/datalad/datalad/builds/220534126>

With the same git-annex version built from source, the tests pass for
me locally.  I haven't tried the conda source, but the Travis logs
indicate that conda falls back to the standalone install for this
version:

  Standalone distribution of git-annex was installed, instead of the
  standard distribution, likely due to package conflicts in the target
  environment.  The standalone distribution may have issues [...]

Given that newer and older conda installs pass on Travis and that I
don't see the failures when building from source, perhaps the failures
are related to the standalone setup issues.  It looks like there is a
more recent version available on conda-forge and judging by the
"alldep" rather than "nodep" in the file name, it should avoid the
standalone fallback.  So try with that.