TST: addurls: Avoid deprecated/racy tempfile.mktemp
Following @mih's efforts from gh-5511, TestAddurls.setup_class() is
one of the remaining spots that use mktemp().  This call could be
switched to mkstemp/NamedTemporaryFile, but setup_class() already
creates a temporary directory with mkdtemp() to use for HTTPPath
content.  Place the JSON file under this directory, moving the content
to a subdirectory.

Note that teardown_class() should have taken care of removing the
temporary JSON file but didn't.  It will now be handled with the
removal of the temporary directory.