RF: use remote.{remote}.datalad-push-default-first config to push first refspec
This is to not incure the run time penalty for other (than github) types
of remotes which do (? at least push by git itself via ssh does) take the
first pushed refspec to be the default branch.

I also removed previously proposed checks on existing branches, since in case
we reconfigure an existing remote (but we do not fetch --prune) we might still
have old remote branches hanging around.  So decided to act solely based on the
config variable.

Moreover, such variable could be set globally if so desired to e.g. overcome
similar issue if detected with other provider without waiting for a fix in the
codebase. We will unset only "local"ly set value