RF: move contains and fullfilled into common_opts + adjust wording
Description was using "report" as the action to describe the effect
on (sub)datasets if those options are given.  Even with original
`subdataset` command "report" is already a stretch since it is not
really a notion in Python -- records or instances of
those (sub)datasets are returned.  Since those options in effect
introduce limiting on which subdatasets are considered for a
command (such as `subdatasets`) I have adjusted wording to use
"considered" instead of "reported".  This IMHO would allow for
reuse of those options in other commands without needing to adjust
`doc` where "reported" would be completely out of place.