BF: interface: Allow positional arguments with underscores
Interface.setup_parser() assumes that argparse replaces all dashes in
the argument's name with underscores. However, argparse leaves dashes
in positional arguments untouched.  This results in an error when
positional arguments specified via Parameter() use names with
underscores (in the Python variant) and dashes (in the command-line
variant).

Trying to work around this by using Parameter(..., dest=...) just
leads to a different error because Interface.setup_parser() doesn't
account for the fact that argparse.add_argument() treats `dest`
differently depending on whether the argument is a positional argument
or an optional one.

Teach setup_parser() to 1) leave positional arguments untouched and 2)
not pass `dest` as a keyword argument to add_argument() when adding a
positional parameter.  Parameter names for positional argument can now
include underscores, either inferred from the parameter name or
explicitly specified via `args`.  If the caller wants the dashed form
to show up in the command-line output, `metavar` can be specified.

Closes #2269.