DOC: download-url: Add examples of path's trailing slash behavior
Since de9d692318 (ENH: download_url: Use trailing separator to signal
directory target, 2019-11-01), the path value is treated as a
directory if a trailing separator is given and as a regular file
otherwise.  --path's description mentions this, but the behavior can
be surprising/confusing, so add examples to highlight the behavior.

Closes #5528.