BF: annexrepo.whereis: Drop success=True assertion for "full" output
When a file doesn't have enough copies, whereis() exits with a
non-zero status, and the "success" field in those records is set to
false.  The "full" output value is the only one that raises an error
in this case because it contains an assertion that, if "success" is
present in the output, it is true.

Drop that assertion for consistent behavior with the other output
modes.

Note that for invalid options, a CommandError is raised upstream.  See
256823e93d (RF: Maintain non-exception behavior of whereis() and
fsck(), 2020-12-08).