RF: annexrepo: Prefer key error over more obscure error
_whereis_json_to_dict() retrieves the "whereis" field with
dict.get("whereis").  It then iterates over this directly, so if the
whereis output from git-annex unexpectedly doesn't have a "whereis"
field, a type error would be signaled.

Change this to a direct key lookup so that the failure is more
informative if the assumption that the "whereis" field is always
present turns out to be wrong.  (There hasn't been an indication that
it is wrong: `for remote in j.get('whereis')` was introduced in 2015
and a type error hasn't yet been reported.)

We could also change this to a more permissive .get("whereis", []),
though that would likely just hide away an underlying issue.  Note
that 'git annex whereis' has an empty "whereis" field when a file
doesn't have any known locations.