BF: update: Abort when ff-only is used with adjusted branches
For adjusted branches, update() goes into an simple 'annex sync' mode
and ff-only is not honored.

Abort in this situation to minimize surprises.  To be less disruptive,
we could add a warning instead, but, if the caller is specifying the
more conservative ff-only, it's probably important to them that it is
honored.