RF: update: Map --merge onto more flexible --how/--how-subds
In preparation for having non-merge actions for updating the working
tree (e.g., "reset --hard"), add --how and --how-subds options.

The motivation behind the separate knobs is that updating the current
dataset and updating the subdatasets aren't inherently coupled
actions.  Allowing them to be specified separately makes it possible
to use update() when the desired result is closer to that of typical
non-datalad approaches like

  git pull --no-recurse-submodules [...]
  git submodule update --recursive [...]

The extra option (--how-subds) doesn't come with any inconvenience
from the caller's perspective because --how-subds falls back to --how,
behaving in the same coupled way that --merge does.