DOC: update: Describe sibling resolution in --sibling, not --merge
The details about sibling resolution are confusingly split across the
descriptions of --sibling and --merge.  And, without internal
knowledge of update(), I think the descriptions even seem to say
conflicting things.

Move all the sibling resolution information into --sibling, and add
details about the interaction with --merge.  This setup also has the
advantage of avoiding repetition when --how and --how-subds are added.