ENH: update: Provide status results for more update actions
update() returns a result for 'git merge TARGET', using status=error
when a CommandError is raised.  This lets the action be fed through
the normal result interface, where it can be controlled by
--on-failure.

Depending on the dataset type and whether it is adjusted, 'git annex
merge' may be called in addition to 'git merge', or 'git annex sync'
may replace the 'git merge' call completely.  Return a result dict for
these actions as well.