ENH: run: Make --dry-run take a value and add 'command' mode
It's likely that the need for more modes will come up in the future,
and, as pointed out by @mih, tacking on optional values later is
awkward/confusing because '--dry-run command arg ...' would take
"command" as the argument to --dry-run.

So, put the current mode under "basic" (inspired by `status
--annex=basic`), and add a mode that shows just the expanded command.

If --dry-run=basic ends up being too verbose, we can add an empty
string value as a shortcut.