TST: gitrepo: Update test for change in git-annex behavior
Back when this test was added in ad0158cb18 (BF(v7): gitrepo: Avoid
adding files to annex, 2019-09-23), 'git add' in a v6+ repo would add
a file to the annex unless annex.largefiles instructed otherwise.
With git-annex 7.20191024, 'git add' doesn't add files to the annex
_unless_ annex.largefiles is set.

Configure annex.largefiles for the test repo so that this test remains
an effective regression test for the GitRepo.add_() change in
ad0158cb18.