ENH: create-sibling: Make default sibling choice more accessible
When --name isn't specified, create-sibling will use the hostname for
an SSH URL and "local" for a local path.  At the default logging
level, however, the caller isn't told what the selected name is.  And,
for scripting purposes, the name can't be easily determined because
it's not available in the result record.

Promote the debug-level message about the default sibling name to an
info-level message, and add a "sibling_name" field to the result
record.

Closes #1986.