Reload config of subdatasets in run-procedure
When run-procedure recursively dives into subdatasets to find a
procedure and its config, those Dataset instances may require a reload
of their ConfigManager. A caller to run-procedure may operate on a
superdataset only may have reloaded its ConfigManager, but doing it
recursively for subdatasets that may contain the procedure in question
would essentially require to reimplement procedure discovery outside.
Moreover here we can ensure that we reload only in subdatasets, that
actually have a match for the procedure.

(Closes #5549)