RF: Move web UI into an extension
The change from the user perspective is that they will have to
install https://github.com/datalad/datalad-deprecated to keep
using the web UI. They will be prompted, if that is not the case.

The tests are kept as-is, but testing web UI related functionality
has been made conditional on the presence of the extension package.

A CI run for the new extension has been added, so even from the test
perspective things remain the same.

Fixes gh-5376