TST: rerun: Avoid direct repo.add() call
test_rerun_explicit() calls repo.add(..., git=True) with an untracked
file, followed by a Dataset.save().  There's no need for this two step
process, and, as of two commits back, the staged file is reset, so the
save() call annexes the file rather than tracking it with git, as
intended.