DOC: common_cfg: Clarify description of inherit-local-origin
clone.configure_origins() correctly handles when
clone.defaultRemoteName is set to value other than "origin".  Given
that, perhaps we'd pick a different name for the
datalad.install.inherit-local-origin option if we were adding it
today, but at this point it's probably not worth changing.

Extend the option's description to mention that
clone.defaultRemoteName is honored.

Note that --reckless is another spot where user-facing documentation
uses "origin", but it's necessarily a referring to a literal remote
name, and it's probably clearer to leave it as is.