TST: Rework tests to not hard code 'origin' as remote name
As of Git 2.30, the caller can configure clone.defaultRemoteName.  A
user doing so won't make the tests fail because we use a custom home,
but ideally we would test with a value other than "origin" to ensure
that DataLad is compatible with a custom value.

Replace all hard-coded "origin" values with a DEFAULT_REMOTE variable
instead to prepare for testing with a custom clone.defaultRemoteName
value.