DOC: Adjust some comments that assume "origin" remote
The previous tweaked spots in the code that assume that git-clone sets
up a remote named "origin" because the name can be configured via
clone.defaultRemoteName.  Adjust some spots in the comments (most
nearby the code changes) to reflect this.

Note that there are still a good number of comments, particularly in
tests, that use "origin", but most of these also use "origin" as the
variable name for the dataset/repo instance, so it seems less
confusing to leave those as is.