TST: no_annex: Avoid racy failure on adjusted branches
test_no_annex() has started to fail frequently on the CrippledFS runs.
I can trigger it with `tools/eval_under_testloopfs` locally.

The underlying issue is that if a file is added to the annex but the
annex.largefiles configuration would send it to git, that file will be
sent to git the next time the clean filter runs on it.  In the context
of this test, the test failure is sporadic because the clean filter
runs on the file ("code/inannex") only when its mtime matches
.git/index's.

I don't know why the above failure has started to happen much more
frequently, and, as mentioned in the git-annex link below, I'm not
sure what expectations here should be.

Avoid the intermittent failure by forcing the mtimes to be the same,
and updating the assertion to expect a dirty tree.  That should lead
to a reliable outcome, and, if git-annex's behavior changes here, a
test failure will let us know.

Re: #5300
https://git-annex.branchable.com/forum/one-off_unlocked_annex_files_that_go_against_large/