ENH(UX,DX): raise RuntimeError if PATH env var is not set
I think it is very reasonable to require for it to be set across all OSes
(IIRC).  Even though we would set PATH to point to the location where we find
git-annex using `distutils.spawn.find_executable`, I think it just buries the
issue of inadequate environment, and IMHO it is better to error out with a
clear message right away.

This should prevent needing to troubleshoot odd execution errors such as
in the case of https://github.com/datalad/datalad/issues/5546 .