TST: run: Add test for unmatched globs and placeholders
If the globbing for an --input or --output value doesn't have any
hits, the pattern is left as is in the results and is accessible
through the {inputs} and {outputs} placeholders.  That's odd and
confusing, but it's not clear to me that the alternative of filtering
them out is better.

For inputs, an upcoming commit will error when a pattern has no hits,
so the specific behavior here doesn't matter too much.  For outputs,
unmatched non-globs need to be available in the placeholders because
they name files that should exist after the command is executed.
Perhaps there should be special treatment of patterns that match
glob.has_magic().

For now, add a test to document this confusing area.