BF: run: Re-glob outputs after command is executed
To unlock output files, run() expands globs prior to running a
command.  For the default invocation, there's no reason to do more
because all changes are saved and the outputs are stored in the commit
message unexpanded.  However, when --explicit or
--expand={both,outputs} is specified, re-globbing is needed to ensure
that an up-to-date list of outputs is captured.

Note that for --explicit, only a subset of cases are affected.  If
there are no matches, GlobbedPaths returns the original pattern.  That
pattern gets passed to GitRepo.get_content_info(), where ls-files
expands the glob.  That is, the unexpected behavior surfaces only
when, before the command is executed, an --output matches some files
and the execution produces additional matching files.

Closes #3448.