TST: globbedpaths: Use a less confusing name for a pattern
test_globbedpaths() points GlobbedPaths to "$test_directory/subdir/",
and then tests globbing _within_ "subdir/".  One of the test cases is
"subdir/", which doesn't have a hit and shows up in the results
because GlobbedPaths (intentionally, though confusingly) returns the
original pattern when there are no hits.

Pick a name that makes it clear that the result is due to there being
no file system matches.