ENH: globbedpaths: Introduce a "strict" mode
When calling glob.glob() with a pattern that doesn't match,
GlobbedPaths chops off the tail and looks for the first sub-pattern
that matches.  It does this so that run() can repeatedly glob and
install the needed subdatasets [1].  expand() returns results that
include these partial patterns.  And if the above doesn't work, the
results include the original pattern, since run() needs the value for
the common case of an --output file that doesn't yet exist.

This behavior doesn't allow run() to figure out what paths are
missing.  Extend GlobbedPaths to expose this information.  Ideally the
new strict mode would be the default mode of operation, but retain the
old behavior as the default to avoid breakage in the unlikely case
that third-party code uses GlobbedPaths.

[1] 2cf27db545 (BF: run: Support globbing in uninstalled datasets,
    2018-09-06)