BF: run: Error when --input pattern doesn't match a file
When a pattern given as --input doesn't match an existing file, run()
gives a warning and proceeds.  This is problematic because a
command-line caller can't control it with --on-failure, and a Python
caller can't easily react to the failure.  Using the GlobbedPaths
functionality added in the previous commit, teach run() to yield an
error result if an --input pattern doesn't have any matches.

The treatment of outputs stays that same, but this opens up the
possibility of later giving a warning/error if an --output pattern
doesn't match an existing file after the command is executed.

Note that using a catch-all "*" as the input in a working tree where
glob.glob("*") returns no matches will now error (as shown in added
test).  That seems acceptable given that this case is likely to be
very rare and that the special-cased "." can be used instead.

Also, note that in order to abort the rest of the run() call, the
caller will need to use --on-failure=stop, just as they would if `get`
fails to download a file [1].  We may want to move to making run()
abort in both cases.

Closes #5583.
[1] https://github.com/datalad/datalad-container/issues/147